refactor(rubric): change criteria weight from normalized decimal to relative integer
- Schema: weight column NUMERIC(5,4) → INTEGER DEFAULT 1
- Entity: BigDecimal weight → Integer weight with @Builder.Default 1
- DTOs: @DecimalMin/Max → @Min(1)/@Max(100), weight now optional
- Publish: remove sum=1.0000 validation, only require ≥1 criterion
- Scoring: CriteriaScoreCalculator already uses weightedSum/totalWeight
- Promptfoo: pass Integer weight directly to assertion map
- Tests: all 7 test files updated for integer weights (371 pass)

+++ server/src/main/java/me/nghlong3004/vqc/api/evaluation/handler/CriteriaScoreCalculator.java
@@ -80,7 +80,7 @@ public class CriteriaScoreCalculator {
             .collect(
                 Collectors.toMap(
                     RubricCriterion::getMetricKey,
-                    c -> c.getWeight() != null ? c.getWeight() : BigDecimal.ONE));
+                    c -> c.getWeight() != null ? BigDecimal.valueOf(c.getWeight()) : BigDecimal.ONE));
 
     BigDecimal weightedSum = BigDecimal.ZERO;
     BigDecimal totalWeight = BigDecimal.ZERO;
